feat(cli): add `cog commit` as duplicate of `coco`
