+++ Cargo.toml
@@ -12,6 +12,7 @@ description = """
 Cocogitto is a set of cli tools for the conventional commit
 and semver specifications.
 """
+autobins = false
 
 [profile.release]
 opt-level = 2
