+++ src/bin/coco.rs
@@ -1,11 +1,9 @@
 #![cfg(not(tarpaulin_include))]
-use std::fmt::Write;
+mod cog_commit;
 
-use cocogitto::{CocoGitto, COMMITS_METADATA};
+use cocogitto::CocoGitto;
 
-use anyhow::{bail, Result};
-use conventional_commit_parser::commit::Separator;
-use itertools::Itertools;
+use anyhow::Result;
 use structopt::clap::{AppSettings, Shell};
 use structopt::StructOpt;
 
@@ -15,56 +13,14 @@ const APP_SETTINGS: &[AppSettings] = &[
     AppSettings::DeriveDisplayOrder,
 ];
 
-const EDIT_TEMPLATE: &str = "# Enter the commit message for your changes.
-# Lines starting with # will be ignored, and empty body/footer are allowed.
-# Once you are done, save the changes and exit the editor.
-# Remove all non-comment lines to abort.
-#
-";
-
-fn prepare_header(typ: &str, message: &str, scope: Option<&str>) -> String {
-    let mut header = typ.to_string();
-
-    if let Some(scope) = scope {
-        write!(&mut header, "({})", scope).unwrap();
-    }
-
-    write!(&mut header, ": {}", message).unwrap();
-
-    header
-}
-
-fn prepare_edit_template(typ: &str, message: &str, scope: Option<&str>, breaking: bool) -> String {
-    let mut template: String = EDIT_TEMPLATE.into();
-    let header = prepare_header(typ, message, scope);
-
-    if breaking {
-        template.push_str("# WARNING: This will be marked as a breaking change!\n");
-    }
-
-    write!(
-        &mut template,
-        "{}\n\n# Message body\n\n\n# Message footer\n# For example, foo: bar\n\n\n",
-        header
-    )
-    .unwrap();
-
-    template
-}
-
-fn commit_types() -> Vec<&'static str> {
-    COMMITS_METADATA
-        .iter()
-        .map(|(commit_type, _)| commit_type.as_ref())
-        .collect()
-}
-
 /// A command line tool to create conventional commits
+///
+/// Deprecated in favor of the `cog commit` utility
 #[derive(StructOpt)]
 #[structopt(name = "Coco", author = "Paul D. <paul.delafosse@protonmail.com>", settings = APP_SETTINGS)]
 struct Cli {
     /// Conventional commit type
-    #[structopt(name = "type", possible_values = &commit_types(), default_value_if("completion", None, "chore"))]
+    #[structopt(name = "type", possible_values = &cog_commit::commit_types(), default_value_if("completion", None, "chore"))]
     typ: String,
 
     /// Commit description
@@ -90,58 +46,20 @@ struct Cli {
 fn main() -> Result<()> {
     let cli = Cli::from_args();
 
+    eprintln!("Warning: `coco` is deprecated, use `cog commit` instead");
+
     if let Some(shell) = cli.completion {
         Cli::clap().gen_completions_to("coco", shell, &mut std::io::stdout());
     } else {
         let cocogitto = CocoGitto::get()?;
 
         let (body, footer, breaking) = if cli.edit {
-            let template = prepare_edit_template(
+            cog_commit::edit_message(
                 &cli.typ,
                 &cli.message,
                 cli.scope.as_deref(),
                 cli.breaking_change,
-            );
-
-            let edited = edit::edit(&template)?;
-
-            if edited.lines().all(|line| {
-                let trimmed = line.trim_start();
-                trimmed.is_empty() || trimmed.starts_with('#')
-            }) {
-                bail!("Aborted commit message edit");
-            }
-
-            let content = edited
-                .lines()
-                .filter(|&line| !line.trim_start().starts_with('#'))
-                .join("\n");
-
-            let cc = conventional_commit_parser::parse(content.trim())?;
-
-            let footers: Option<String> = if cc.footers.is_empty() {
-                None
-            } else {
-                Some(
-                    cc.footers
-                        .iter()
-                        .map(|footer| {
-                            let separator = match footer.token_separator {
-                                Separator::Colon => ": ",
-                                Separator::Hash => " #",
-                                Separator::ColonWithNewLine => " \n",
-                            };
-                            format!("{}{}{}", footer.token, separator, footer.content)
-                        })
-                        .join("\n"),
-                )
-            };
-
-            (
-                cc.body.map(|s| s.trim().to_string()),
-                footers,
-                cc.is_breaking_change || cli.breaking_change,
-            )
+            )?
         } else {
             (None, None, cli.breaking_change)
         };
